requirements.txt
matplotlib==3.5.3
numpy==1.21.5
pandas==1.3.5
plotly==5.9.0
PyYAML==6.0
scikit_learn==1.0.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
umap_learn==0.5.3
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
